api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
minio==7.1.17
psycopg2-binary==2.9.7
redis==4.6.0
